Refactor StatsPageLoader folder
