+++ client/my-sites/stats/pages/purchase/controller.tsx
@@ -4,7 +4,7 @@ import type { Context } from '@automattic/calypso-router';
 
 function purchase( context: Context, next: () => void ) {
 	context.primary = (
-		// DO NOT WRAP WITH <LoadStatsPage /> or you will get an infinite loop
+		// DO NOT WRAP WITH <StatsPageLoader /> or you will get an infinite loop
 		<AsyncLoad
 			require="calypso/my-sites/stats/pages/purchase"
 			placeholder={ PageLoading }
